Initial integration of MapCompose map
This code is the initial integration of the MapCompose open street map
UI. The soundscape service connection has been moved out into a separate
class and then a MapCompose based UI has been added to the Home screen.
The service now uses lastLocation to give a much quicker 'lock' on the
current position.

An HTTP cache has also been initiated - this likely affects all HTTP and not
just the Tiles for open street map. As a result it may be required to mark
Tile HTTP requests with a cache-control header if we don't want them to go
through the cache?

The API key for thunderforest is stored in local.properties as
tileProviderApiKey so must be configured for each developer i.e.

tileProviderApiKey=xxXXxxXXxxXXxxXXxxXXxxXXxxXXxxXX

with an appropriate key. For builds, there's a new GitHub secret
TILE_PROVIDER_API_KEY that needs setup.