Add monkey testing of app to nightly builds
Monkey is a random key presser function within adb. This change adds
the installation and running of our app on emulators for each of our
tested API versions. It then runs monkey for 2000 events to give a
little more coverage than running just instrumented tests.
The app has been altered so that a DISABLE_EXIT boolean flag can be
passed into the MainActivity intent. That stops the app from exiting
when the Exit button is pressed.