Add initial Earcon support to audio engine
The iOS Soundscape has lots of audio 'earcons' which are used to indicate
state changes happening in the app e.g. going online/offline from network,
or arriving at a beacon. This initial support allows for a non-positioned
audio file playback which is queued along with the text to speech.