UI additions to allow waypoints to be added to routes and reordered
This change allows the selection of Markers to be added to Routes. It's
not pretty from a UI point of view, but I think the code is at least
reasonably structured.

ReorderableLocationList adds a list of LocationItems which can be
dragged to reorder. Currently the whole item can be clicked to drag, but
really we just want a drag handle target on the end.

Addition of ObjectIds to MarkerData and RouteData for unique identification.