Messing with adding water from MVT
The names appear at specific points on the map. They share the OSM with the
polygon. However, this makes it difficult to get a name for a piece of water
that you only have the edge of. This might need some changes to the protomaps
output to add name to every lake, pond and waterway.