Move to using CheapRuler for measurements in place of full geodesic
CheapRuler is a JavaScript library from MapBox which is very accurate
when used over city sized distances, and much more efficient than full
geodesic calculations. The FeatureTree searches and MapMatch filtering
has massively increased the number of distance calculations performed,
and so this is a good way of regaining some of the performance lost.