Remove foreign HashMap from Feature
We don't ingest Features from GeoJSON so there's no need to have foreign.
The original Soundscape backend server had GeoJSON with foreign entries,
mainly so that it could have a list of OSM ids for intersections. We
calculate intersections within the app and have a Way graph which
describes them and so we don't need the OSM ids list.