Introduce MvtFeature
The aim of this is to reduce the amount of HashMap searching that has to
be done for the commonly used properties. Currently that list is:

	osmId
	name
	class
	subclass
	feature_type
	feature_value

The final two we derive, so we can probably get rid of those altogether as
they were done for backwards compatibility with the soundscape-backend server.

The GeoJSON export now writes the values back into the properties so that it
should output the same data as before. We need to check if that's always what
we want.