Introduce MvtFeature
The aim of this is to reduce the amount of HashMap searching that has to
be done for the commonly used properties. Currently that list is:

	osmId
	name
	class
	subclass
	feature_type
	feature_value

The final two we derive, so we can probably get rid of those altogether as
they were done for backwards compatibility with the soundscape-backend server.

We need to add in GeoJSON export so that it puts these values back into
properties - writing out GeoJSON is purely for debug.