HACK - Adding of housenumber layer parsing and some search tests
Tests for both forward and reverse geocoding including use of a library
(AndroidAddressFormatter) to present the addresses in the correct format
for their location. That is now used for both Photon and Android output
and the results tend to be fairly close.

During tile parsing we now parse the housenumber layer into a number of
FeatureTrees, one for each named road. This makes it much easier to reverse
geocode when we're map matched - or indeed just assuming the nearest road.
The Geocoder Test now gives the most accurate results with LocalDecoder for
the first test where the location is on a corner between two streets. When
there's a map matched Way the LocalDecoder uses it, whereas the other two
are performing a search without any extra knowledge.

There's also a new StreetDescription class. This follows a named street and
creates linear maps of street numbers and point of interest. This can then
be looked up to get an interpolated street address and a description of the
location relative to intersections and POI.