Fix name confection and ensure properties always exists when calling set
getName was mistakenly setting the member variable name the first time it was
called. This change stops that from happening as we want to always force a
re-confection rather than caching the result.
Also, there were various places where `feature.properties?.set()` was being
called where there was a risk of properties being null, though the code was
assuming it was not. Although it was safe to call, the `set` gets lost. The
remaining changes resolve this in a number of ways:

 1. Addition of `setProperty` function in `MvtFeature` which will initialise
    properties if it is null.
 2. Use of `.apply()` when creating the properties from fresh to simplify the
    logic. This is much more idiomatic kotlin.
 3. Use of `.also()` to create properties if it doesn't exist and set the values
    in it. Also more idiomatic kotlin.