Add reference results checking for Mvt testCallouts unit test
The testCallouts unit test generates some GeoJSON for the input GPX file along
with a text file of the callouts made. This change adds some reference text
files so that the test will fail if there are any changes in output. The references
were made with v0.1.6. Any change in callouts and the test will fail allowing us to
verify whether or not the changes are expected or unexpected.